Barrister Sami Rahman speak's Direct Access Risk
The decision to allow patients to access certain health fields through direct access was made according to the highest standards of care, and it was based on a comprehensive review of the available evidence. One of the key findings of the review was a literature review that assessed the effects of various interventions on patient safety.
However, a public seminar was much needed to assess the situations and if the pros outweighed the cons.
A seminar was announced for 23 September, 2015 to collect a panel of different barristers to sit together and discuss the applications of Direct Access and whether there could be more improvements.
The much awaited and much needed opinion came from Barrister Sami Rahman, a property and housing Barrister who practices in a range of fields. We have learned so much from him in the fast, that it felt like a relief to be speaking in favor of this controversial topic.
Sign up for our newsletter to access the detailed report OR watch this complete recorded video of the seminar session on youTube to know for yourself:
https://www.youtube.com/watch?v=R74ypcwqZcU


check out more articles : 
•	Barrister Sami Rehman Slams an Exploiter Suit – in SUPREME COURT!
•	 Barrister Sami Rahman on a spree to end workplace discrimination!
•	 Is whistleblowing a crime in UK? Barrister Sami Rahman puts forward a question.